enumerate all frequency conditions -- count and percent
